config: add config.Load function
