+++ cli/server/command.go
@@ -70,7 +70,7 @@ Examples:
 	var logger log.Logger
 
 	cmd.PreRun = func(_ *cobra.Command, _ []string) {
-		if err := loadConf.Load(conf); err != nil {
+		if err := pikoconfig.Load(conf, loadConf.Path, loadConf.ExpandEnv); err != nil {
 			fmt.Println(err.Error())
 			os.Exit(1)
 		}
