+++ cli/test/workload/traffic/command.go
@@ -61,7 +61,7 @@ Examples:
 	var logger log.Logger
 
 	cmd.PersistentPreRun = func(_ *cobra.Command, _ []string) {
-		if err := loadConf.Load(conf); err != nil {
+		if err := pikoconfig.Load(conf, loadConf.Path, loadConf.ExpandEnv); err != nil {
 			fmt.Println(err.Error())
 			os.Exit(1)
 		}
